split off create / verify pcs proof into new file
