+++ src/core/mod.rs
@@ -11,11 +11,13 @@ pub mod interactive_protocol;
 pub mod messages;
 pub mod setup;
 pub mod state;
-pub mod evaluate;
 pub mod commit;
+pub mod evaluate;
+pub mod pcs;
 
-pub use evaluate::*;
 pub use commit::*;
+pub use evaluate::*;
+pub use pcs::*;
 
 pub use builder::*;
 pub use inner_product::*;
